requirements.txt
flask==3.0.2
requests==2.31.0
markdown==3.5.2
gTTS==2.5.1
python-dotenv==1.0.1
openai==1.12.0
google-generativeai==0.3.2
huggingface-hub==0.20.3 